backend/requirements.txt
fastapi==0.111.0
pymongo==4.6.3
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9